refactor: verseCardDisplay feature로 분리
