+++ src/features/drilling/index.tsx
@@ -1,9 +1,6 @@
-import CardDisplay from 'src/features/drilling/components/verseDisplay';
-import { ComposedBoundary } from '@/lib/error/ComposedBoundary';
-import ErrorMessage from '@/lib/error/ErrorMessage';
-import Loader from '@/shared/ui/Loader';
 import BibleVersionSelect from '@features/bibleVersionSelect';
 import CardHideOptionSelect from '@features/cardHideOptionSelect';
+import VerseCardDisplay from '@features/verseCardDisplay';
 
 function Drilling() {
   return (
@@ -12,18 +9,7 @@ function Drilling() {
         <BibleVersionSelect />
         <CardHideOptionSelect />
       </div>
-      <ComposedBoundary
-        fallbackRender={({ resetErrorBoundary, error }) => (
-          <ErrorMessage
-            className='my-8 h-[400px] rounded-3xl bg-neutral-100 px-9 py-6 align-middle mobile:h-[200px]'
-            resetErrorBoundary={resetErrorBoundary}
-            error={error}
-          />
-        )}
-        suspenseFallback={<Loader size='lg' className='my-[120px]' />}
-      >
-        <CardDisplay />
-      </ComposedBoundary>
+      <VerseCardDisplay />
     </div>
   );
 }
