+++ src/features/verseCardDisplay/components/cardSwiper/index.test.tsx
