+++ src/features/verseCardDisplay/components/verseCard/type.ts
